Solved a minor bug on MTF
