--- tgen/graficar.py
@@ -1,101 +1,0 @@
-#!/home/adriano/Escritorio/TFG/venv/bin/python3
-import math
-import numpy as np
-import matplotlib.pyplot as plt
-import activity_data as act
-import recurrence_plots as rec
-import GAF as gaf
-import GAF2 as gaf2
-import REC  as rec
-from PIL import Image
-import cv2
-
-
-
-def main():
-     Tipo="GAF"
-     data_name="WISDM"
-     data_folder="/home/adriano/Escritorio/TFG/data/WISDM/"
-     #voy a  obtener el maximo de todo el data set.
-     X_train, y_train, sj_train = act.load_numpy_datasets(data_name, data_folder, USE_RECONSTRUCTED_DATA=False)
-     print("X_train", X_train.shape, "y_train", y_train.shape, "sj_train", sj_train.shape)
-     tipos=["GAF","GAF2","RP","RP2","MTF","MTF2"]
-     MAX=np.max(X_train)
-     MIN=np.min(X_train)
-     a=0
-     w = X_train[a]
-     sj = sj_train[a][0]
-     w_y = y_train[a]
-     w_y_no_cat = np.argmax(w_y)
-     print(w.shape)
-     
-     dictionary=dict()
-     for k in range(0,8163):
-         l=X_train[k]
-         #A=np.max(l[:,0])
-         maximos=[np.max(l[:,0]),np.max(l[:,1]),np.max(l[:,2])]
-         minimos=[np.min(l[:,0]),np.min(l[:,1]),np.min(l[:,2])]
-         dictionary[k]=[maximos,minimos]
-
-     
-     img = SavevarGAF_XYZ(w, sj, a, "x", normalized = 1, path=f"./", TIME_STEPS=129)
-     imagen = cv2.imread("./1600x0gasf.png")  
-     imagen = cv2.cvtColor(imagen, cv2.COLOR_BGR2RGB)
-     print("Image shape",imagen.shape)
-     
-     rp=Reconstruct_GAF(imagen,dictionary,a)
-     # Configurar el estilo de los gráficos
-    
-     dim=0
-     
-    # Configurar el estilo de los gráficos
-     plt.style.use("ggplot")  
-
-    # Gráfico original
-     plt.figure(figsize=(10, 6))
-     plt.plot(w[:, dim], marker='o', color='blue')
-     plt.title('Original', fontsize=18,fontweight="bold")
-     plt.xlabel("Tiempo", fontsize=12)
-     plt.ylabel('Índice X', fontsize=12)
-     plt.grid(True)
-     plt.tight_layout()
-     plt.savefig('original.png', bbox_inches='tight', pad_inches=0)
-     plt.clf()
-
-    # Gráfico reconstrucción
-     plt.figure(figsize=(10, 6))
-     plt.plot(rp[dim], marker='o', color='green')
-     plt.title('Reconstrucción', fontsize=18,fontweight="bold")
-     plt.xlabel('Tiempo', fontsize=12)
-     plt.ylabel('Índice X', fontsize=12)
-     plt.grid(True)
-     plt.tight_layout()
-     plt.savefig('reconstruccion.png', bbox_inches='tight', pad_inches=0)
-     plt.clf()
-
-    # Gráfico comparativa
-     plt.figure(figsize=(10, 6))
-     plt.plot(w[:, dim], marker='o', label='Original', color='blue')
-     plt.plot(rp[dim], marker='o', label='Reconstrucción', color='green')
-     plt.title('Comparativa', fontsize=18,fontweight="bold")
-     plt.xlabel("Tiempo", fontsize=12)
-     plt.ylabel('Índice X', fontsize=12)
-     plt.legend(fontsize=12)
-     plt.grid(True)
-     plt.tight_layout()
-     plt.savefig('Comparativa.png', bbox_inches='tight', pad_inches=0)
-     plt.clf()
-     
-     f=np.array(w[:,dim])
-     f=f[:]
-     print(f.shape)
-     error_absoluto, error_relativo = calcular_errores(f, rp[dim])
-     #d = dtw.distance_fast(f, rp[1], use_pruning=True)
-     print(f"Error Absoluto Promedio: {error_absoluto}")
-     print(f"Error Relativo Promedio: {error_relativo}")
-     #print(f"Error DTW: {d}")
-     print(f"Coeficiente de correlación: {np.corrcoef(f, rp[dim])[0,1]}")
-     
-
-if __name__ == '__main__':
-    main()
